rustdb::common

Module utils

Utilities for rustdb

Functions

calculate_hash

Calculates a hash for the given value

calculate_max_record_size

Calculates the maximum record size on a page

calculate_optimal_btree_order

Calculates the optimal B+ tree order for the given key size

calculate_optimal_hash_table_size

Calculates the optimal hash table size for the given number of elements

calculate_optimal_page_size

Calculates the optimal page size for the given data size

calculate_page_header_size

Calculates the page header size based on the page size

calculate_pages_needed

Calculates the number of pages needed to store data

can_fit_record_on_page

Checks if a record can fit on a page

format_bytes

Formats bytes size into a readable format

format_duration

Formats time in seconds into a readable format

is_power_of_two

Checks if a number is a power of 2

is_valid_column_name

Checks if the column name is valid

is_valid_index_name

Checks if the index name is valid

is_valid_page_size

Checks if the page size is supported

is_valid_table_name

Checks if the table name is valid

next_power_of_two

Calculates the nearest power of 2 greater than or equal to the given number

prev_power_of_two

Calculates the nearest power of 2 less than or equal to the given number

should_expand_hash_table

Checks if a hash table should be expanded

should_merge_pages

Checks if pages should be merged

should_shrink_hash_table

Checks if a hash table should be shrunk

should_split_page

Checks if a page should be split