rustdb::storage::page

Struct Page

pub struct Page {
    pub header: PageHeader,
    pub slots: Vec<RecordSlot>,
    pub data: Vec<u8>,
    pub free_space_map: Vec<bool>,
}

Page structure

Fields
header: PageHeader

Page header

slots: Vec<RecordSlot>

Record slots

data: Vec<u8>

Page data

free_space_map: Vec<bool>

Free space map

Implementations

impl Page

pub fn new(page_id: PageId) -> Self

Creates a new empty data page

pub fn new_with_type(page_id: PageId, page_type: PageType) -> Self

Creates a new empty page with the specified type

pub fn from_bytes(bytes: &[u8]) -> Result<Self>

Creates a page from bytes (deserialization) Detects format: slotted (magic RP) or legacy bincode

fn from_bytes_bincode(bytes: &[u8]) -> Result<Self>

Deserializes from legacy bincode format

fn from_bytes_slotted(bytes: &[u8]) -> Result<Self>

Deserializes from slotted binary format

pub fn from_bytes_with_id(bytes: &[u8], page_id: PageId) -> Result<Self>

Creates a page from bytes with specified page_id

pub fn to_bytes(&self) -> Result<Vec<u8>>

Serializes the page to bytes (uses slotted format for compactness)

fn to_bytes_slotted(&self) -> Result<Vec<u8>>

Serializes to slotted binary format (compact, fixed layout)

pub fn add_record(&mut self, record_data: &[u8], record_id: u64) -> Result<u32>

Adds a record to the page

pub fn delete_record(&mut self, record_id: u64) -> Result<bool>

Deletes a record by ID

pub fn get_record(&self, record_id: u64) -> Option<&[u8]>

Gets a record by ID

pub fn update_record(&mut self, record_id: u64, new_data: &[u8]) -> Result<bool>

Updates a record

fn find_free_space(&self, size: usize) -> Result<usize>

Finds free space for a record of the specified size

fn update_free_space_map(&mut self, start: usize, end: usize, is_free: bool)

Updates the free space map

pub fn has_free_space(&self, size: usize) -> bool

Checks if there is free space for a record of the specified size

pub fn free_space(&self) -> u32

Returns the amount of free space

pub fn record_count(&self) -> u32

Returns the number of records

pub fn is_empty(&self) -> bool

Checks if the page is empty

pub fn is_full(&self) -> bool

Checks if the page is full

pub fn clear(&mut self) -> Result<()>

Clears the page

pub fn get_fill_factor(&self) -> f64

Gets the page fill factor (0.0 - 1.0)

pub fn get_record_count(&self) -> u32

Gets the number of records

pub fn get_free_space(&self) -> u32

Gets the free space size

pub fn needs_defragmentation(&self) -> bool

Checks if defragmentation is needed

pub fn scan_records(&self) -> Result<Vec<(u32, Vec<u8>)>>

Scans all records on the page

pub fn get_all_records(&self) -> Result<Vec<Vec<u8>>>

Gets all records on the page

pub fn defragment(&mut self) -> Result<()>

Performs page defragmentation

pub fn update_record_by_offset( &mut self, offset: u32, new_data: &[u8], ) -> Result<()>

Updates a record by offset

pub fn delete_record_by_offset(&mut self, offset: u32) -> Result<()>

Deletes a record by offset